+++ tests/util.py
@@ -1,6 +1,9 @@
 # -*- coding: utf-8 -*-
-# (c) 2018 The mqttwarn developers
+# (c) 2018-2021 The mqttwarn developers
 import time
+from dataclasses import dataclass
+from typing import Dict, Union, List
+
 from paho.mqtt.client import MQTTMessage
 
 from mqttwarn.configuration import load_configuration
@@ -38,3 +41,19 @@ def send_message(topic=None, payload=None):
 
     # Give the machinery some time to process the message
     time.sleep(0.05)
+
+
+@dataclass
+class ProcessorItem:
+    """
+    A surrogate processor item for feeding into service handlers.
+    """
+
+    service: str = None
+    target: str = None
+    config: Dict = None
+    addrs: List[str] = None
+    topic: str = None
+    title: str = None
+    message: Union[str, bytes] = None
+    data: Dict = None